+++ stix_shifter/stix_translation/src/modules/bigfix/bigfix_translator.py
@@ -1,6 +1,6 @@
 from ..base.base_translator import BaseTranslator
 from .stix_to_bigfix import StixToRelevanceQuery
-from .bigfix_to_stix.big_to_stix import BigFixToStix
+from ...json_to_stix.json_to_stix import JSONToStix
 
 from os import path
 
@@ -12,5 +12,5 @@ class Translator(BaseTranslator):
             path.join(basepath, "json", "to_stix_map.json"))
         
         self.mapping_filepath = filepath
-        self.result_translator = BigFixToStix(filepath)
+        self.result_translator = JSONToStix(filepath)
         self.query_translator = StixToRelevanceQuery()
